lesson_10/requirements.txt
selenium==4.16.0
pytest==8.4.1
webdriver-manager==4.0.1
allure-pytest==2.13.0